requirements.txt
flask==1.0.2
pillow==5.1.0
requests==2.18.4
